% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nars_links}
\alias{nars_links}
\title{NARS Data Links}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 268 rows and 6 columns.
}
\usage{
nars_links
}
\description{
A dataset containing data links for NARS data
}
\details{
format A data frame with 268 rows and 6 variables:
\describe{
\item{survey_type}{NARS survey type}
\item{survey_year}{Survey year}
\item{indicator}{Indicator from survey}
\item{data}{Description of indicator data}
\item{data_url}{Link to data}
\item{metadata}{Link to metadata}
}
}
\keyword{datasets}
