% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_nars.R
\name{scrape_nars}
\alias{scrape_nars}
\title{Scrape NARS data}
\usage{
scrape_nars(survey, year, parameter, type = "data")
}
\arguments{
\item{survey}{Survey name}

\item{year}{Survey year}

\item{parameter}{Survey parameter}

\item{type}{Data type; either data or metadata}
}
\description{
Scrape NARS data
}
